+++ Godeps/_workspace/src/github.com/vishvananda/netlink/route_unspecified.go
@@ -1,0 +1,7 @@
+// +build !linux
+
+package netlink
+
+func (r *Route) ListFlags() []string {
+	return []string{}
+}
